+++ src/array/mod.rs
@@ -1,7 +1,7 @@
 mod r#async;
+mod builder;
 mod chunk_grid;
 mod chunk_key_encoding;
-mod create;
 mod selection;
 mod shared;
 mod sync;
